requirements.txt
streamlit
python-dotenv
pymupdf
pytesseract
Pillow
langchain
langchain-community
langchain-huggingface
langchain-core
faiss-cpu
sentence-transformers
transformers
torch
huggingface-hub
pandas
numpy